Create designer.md
Created a vacancy for the designer. 
@atatarenko, could you please review it? 
And could I please ask you to pay more attention to the "what will you do unit"?  It was created from my head, so there may be more things to change:)

Also, there is a section with the questions to filter the candidates. What questions would you like to ask?